the great edward renaming of 2018
